backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv==1.0.0
channels
channels-redis
twilio==8.12.0
django-jazzmin
jdatetime